refactored folder structure top bloc type
