+++ lib/main.dart
@@ -1,11 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:nxt_gen_guidance/colors.dart';
-import 'package:nxt_gen_guidance/pages/home_page.dart';
-import 'package:nxt_gen_guidance/pages/login_page.dart';
-import 'package:nxt_gen_guidance/pages/main_page.dart';
+import 'package:nxt_gen_guidance/login/view/login_page.dart';
 
 void main() {
-  runApp(MyApp());
+  runApp(const MyApp());
 }
 
 class MyApp extends StatelessWidget {
@@ -14,7 +12,7 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      home: LoginPage(),
+      home: const LoginPage(),
       theme: ThemeData(useMaterial3: true, colorSchemeSeed: kBlueColor),
     );
   }
